# docs/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DeNoise application API",
    "description": "Authenticated operator API (docs/06-api-contract.md). Session cookie for the UI or \u0060Authorization: Bearer ah_pat_\u003CkeyId\u003E.\u003Csecret\u003E\u0060 for API clients.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/integrations": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "operationId": "ListIntegrations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "post": {
        "tags": [
          "Integrations"
        ],
        "operationId": "CreateIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/integrations/{id}": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "operationId": "GetIntegration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "put": {
        "tags": [
          "Integrations"
        ],
        "operationId": "UpdateIntegration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
...
    "/api/v1/integrations/{id}/failures": {
      "get": {
        "tags": [
          "Mappings"
        ],
        "operationId": "ListMappingFailures",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "all",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MappingFailureDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/integrations/{id}/failures/{failureId}/dismiss": {
      "post": {
        "tags": [
          "Mappings"
        ],
        "operationId": "DismissMappingFailure",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "failureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/replay": {
      "post": {
        "tags": [
          "Replay"
        ],
        "operationId": "StartReplay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayStatus"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/replay/{jobId}": {
      "get": {
        "tags": [
          "Replay"
        ],
        "operationId": "GetReplay",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIntegrationRequest": {
        "required": [
          "name",
          "type",
          "accessScope",
          "ownerTeamId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "accessScope": {
            "type": "string"
          },
          "ownerTeamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "hmacSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "hmac": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HmacSettings"
              }
            ]
          },
          "capabilities": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverage": {
            "type": [
              "null",
              "string"
            ]
          },
          "profileDefaults": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "IntegrationCreatedResponse": {
        "required": [
          "integration",
          "ingestPath",
          "ingestToken"
        ],
        "type": "object",
        "properties": {
          "integration": {
            "$ref": "#/components/schemas/IntegrationSummary"
          },
          "ingestPath": {
            "type": "string"
          },
          "ingestToken": {
            "type": "string"
          }
        }
      },
...
      "IntegrationSummary": {
        "required": [
          "id",
          "name",
          "type",
          "accessScope",
          "ownerTeamId",
          "ingestKeyId",
          "active",
          "shadow",
          "version",
          "activatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "accessScope": {
            "type": "string"
          },
          "ownerTeamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "ingestKeyId": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "shadow": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "activatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "hmacConfigured": {
            "type": "boolean",
            "default": false
          },
          "hmac": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HmacSettings"
              }
            ]
          },
          "capabilities": {
            "type": "string",
            "default": "{}"
          },
          "coverage": {
            "type": "string",
            "default": "{}"
          },
          "profileDefaults": {
            "type": "string",
            "default": "{}"
          },
          "ipAllowList": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "stateQuery": {
            "type": "boolean",
            "default": false
          },
          "atlasGroupId": {
            "type": [
              "null",
              "string"
            ]
          },
          "atlasBaseUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "atlasCredentials": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "JsonElement": {},
...
      "MappingFailureDto": {
        "required": [
          "id",
          "eventId",
          "receivedAt",
          "mappingVersion",
          "error",
          "field",
          "quarantined",
          "resolvedAt",
          "rawExcerpt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "mappingVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "field": {
            "type": [
              "null",
              "string"
            ]
          },
          "quarantined": {
            "type": "boolean"
          },
          "resolvedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rawExcerpt": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReplayStatus": {
        "required": [
          "jobId",
          "status",
          "mode",
          "integrationId",
          "attempts",
          "lastError",
          "createdAt",
          "updatedAt",
          "result"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "integrationId": {
            "type": "string",
            "format": "uuid"
          },
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "lastError": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonElement"
              }
            ]
          }
        }
      },
...
      "StartReplayRequest": {
        "required": [
          "mode",
          "integrationId"
        ],
        "type": "object",
        "properties": {
          "mode": {
            "type": "string"
          },
          "integrationId": {
            "type": "string",
            "format": "uuid"
          },
          "eventIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "from": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "to": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "mappingVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "default": 5000
          }
        }
      },
...
      "UpdateIntegrationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "accessScope": {
            "type": [
              "null",
              "string"
            ]
          },
          "ownerTeamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clearOwnerTeam": {
            "type": "boolean",
            "default": false
          },
          "capabilities": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverage": {
            "type": [
              "null",
              "string"
            ]
          },
          "profileDefaults": {
            "type": [
              "null",
              "string"
            ]
          },
          "ipAllowList": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "hmacSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "hmac": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HmacSettings"
              }
            ]
          },
          "clearHmac": {
            "type": "boolean",
            "default": false
          },
          "atlasApi": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AtlasApiCredentials"
              }
            ]
          },
          "active": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "shadow": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "description": "Server-side session cookie set by POST /auth/login.",
        "name": "denoise_session",
        "in": "cookie"
      },
      "pat": {
        "type": "http",
        "description": "Personal access token ah_pat_\u003CkeyId\u003E.\u003Csecret\u003E.",
        "scheme": "bearer"
      }
    }
  },
...
}